+++ lib/release_utility.dart
@@ -1,0 +1,4 @@
+/// Release Updater Utility Library.
+library release_updater.utility;
+
+export 'src/windows_pe_file.dart';