+++ test/release_packer_test.dart
@@ -63,6 +63,81 @@ void main() {
         expect(authorization.username, equals('userX'));
         expect(authorization.password, equals('pass123'));
       }
+
+      {
+        var cmd = ReleasePackerCommand.from({
+          'url': {
+            'url': 'http://foo/bar',
+            'authorization': {'user': 'userX', 'pass': 'pass123'},
+            'parameters': {'a': 123},
+            'body': 'Data'
+          }
+        });
+
+        expect(cmd, isA<ReleasePackerCommandURL>());
+
+        var cmdURL = cmd as ReleasePackerCommandURL;
+
+        expect(cmdURL.url, equals('http://foo/bar'));
+
+        var authorization = cmdURL.authorization as BasicCredential;
+        expect(authorization.username, equals('userX'));
+        expect(authorization.password, equals('pass123'));
+      }
+
+      {
+        var cmd = ReleasePackerCommand.from('dart_pub_get');
+
+        expect(cmd, isA<ReleasePackerDartPubGet>());
+      }
+
+      {
+        var cmd =
+            ReleasePackerCommand.from({'dart_compile_exe': 'bin/foo.dart'});
+
+        expect(cmd, isA<ReleasePackerDartCompileExe>());
+
+        var cmdWinGUI = cmd as ReleasePackerDartCompileExe;
+
+        expect(cmdWinGUI.args, equals(['exe', 'bin/foo.dart']));
+      }
+
+      {
+        var cmd = ReleasePackerCommand.from({'windows_gui': 'bin/foo.exe'});
+
+        expect(cmd, isA<ReleasePackerWindowsSubsystemCommand>());
+
+        var cmdWinGUI = cmd as ReleasePackerWindowsSubsystemCommand;
+
+        expect(cmdWinGUI.args, equals(['--windows-gui', 'bin/foo.exe']));
+      }
+
+      {
+        var cmd = ReleasePackerCommand.from(['dart', 'pub', 'get']);
+
+        expect(cmd, isA<ReleasePackerDartPubGet>());
+      }
+
+      {
+        var cmd = ReleasePackerCommand.from(
+            ['dart', 'compile', 'exe', 'bin/foo.dart']);
+
+        expect(cmd, isA<ReleasePackerDartCompileExe>());
+
+        var cmdDartExe = cmd as ReleasePackerDartCompileExe;
+
+        expect(cmdDartExe.args, equals(['exe', 'bin/foo.dart']));
+      }
+
+      {
+        var cmd = ReleasePackerCommand.from(['windows_gui', 'bin/foo.exe']);
+
+        expect(cmd, isA<ReleasePackerWindowsSubsystemCommand>());
+
+        var cmdWinGUI = cmd as ReleasePackerWindowsSubsystemCommand;
+
+        expect(cmdWinGUI.args, equals(['--windows-gui', 'bin/foo.exe']));
+      }
     });
 
     test('buildFromDirectory', () async {
@@ -81,7 +156,7 @@ void main() {
       expect(releasePacker.version.toString(), equals(ReleaseUpdater.VERSION));
 
       var prepareCommands = releasePacker.prepareCommands!;
-      expect(prepareCommands.length, equals(3));
+      expect(prepareCommands.length, equals(4));
 
       {
         expect(prepareCommands[0], isA<ReleasePackerDartPubGet>());
@@ -90,11 +165,20 @@ void main() {
         expect((prepareCommands[1] as ReleasePackerDartCompileExe).args,
             equals(['exe', 'bin/foo.dart']));
 
-        expect(prepareCommands[2], isA<ReleasePackerProcessCommand>());
-        expect((prepareCommands[2] as ReleasePackerProcessCommand).command,
+        expect(prepareCommands[2], isA<ReleasePackerWindowsSubsystemCommand>());
+        expect(
+            (prepareCommands[2] as ReleasePackerWindowsSubsystemCommand)
+                .command,
+            equals('release_utility'));
+        expect(
+            (prepareCommands[2] as ReleasePackerWindowsSubsystemCommand).args,
+            equals(['--windows-gui', 'bin/foo.exe']));
+
+        expect(prepareCommands[3], isA<ReleasePackerProcessCommand>());
+        expect((prepareCommands[3] as ReleasePackerProcessCommand).command,
             equals('bin/foo.exe'));
         expect(
-            (prepareCommands[2] as ReleasePackerProcessCommand).stdoutFilePath,
+            (prepareCommands[3] as ReleasePackerProcessCommand).stdoutFilePath,
             equals('foo.out'));
       }